AzizFiverrXnarx-main/requirements.txt
mysql-connector-python==8.4.0
pandas==2.1.4
Scrapy==2.11.2
selenium_driverless==1.9.3.1
requests==2.31.0
playwright==1.43.0
deep-translator
openpyxl